Exception Handler

+++ src/main/java/DSCommerce/services/ProductService.java
@@ -1,14 +1,19 @@
 package DSCommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import DSCommerce.DTO.ProductDTO;
 import DSCommerce.entities.Product;
 import DSCommerce.repositories.ProductRepository;
+import DSCommerce.services.exceptions.DatabaseException;
+import DSCommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,7 +23,9 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+
+		Product product = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("resource not found"));
 		return new ProductDTO(product);
 	}
 
@@ -39,16 +46,30 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		catch(EntityNotFoundException e) {
+			 throw new ResourceNotFoundException("resource not found");
+		}
+		
 
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("resource not found");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("referential integrity failure");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
